refactor: extract useClassColors composable; fix pre-existing noEdgesResponse unused import
